Delete employee refactored

+++ src/main/java/files/kinoxp/service/MovieScreeningService.java
@@ -1,8 +1,11 @@
 package files.kinoxp.service;
 
+import files.kinoxp.model.Movie;
 import files.kinoxp.model.MovieScreening;
+import files.kinoxp.repository.MovieRepository;
 import files.kinoxp.repository.MovieScreeningRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +16,7 @@ public class MovieScreeningService {
 
     @Autowired
     MovieScreeningRepository movieScreeningRepository;
+    private MovieRepository movieRepository;
 
     public MovieScreening createMovieScreening(MovieScreening movieScreening) {
         return movieScreeningRepository.save(movieScreening);
@@ -22,8 +26,14 @@ public class MovieScreeningService {
         return movieScreeningRepository.findAll();
     }
 
-    public void deleteMovieScreening(int id) {
-        movieScreeningRepository.deleteById(id);
+    public ResponseEntity<HttpStatus> deleteMovieScreening(int id) {
+        if (movieScreeningRepository.findById(id).isPresent()) {
+            movieScreeningRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND, HttpStatus.NOT_FOUND);
+        }
+
     }
 
     public ResponseEntity<MovieScreening> updateMovieScreening(MovieScreening movieScreening, int id) {
